fix: move common functions to deploy module
