+++ bench/Cargo.toml
@@ -9,14 +9,12 @@ publish = false
 
 [dependencies]
 arrow = { git = "https://github.com/flock-lab/arrow-rs", branch = "flock", features = [ "simd" ] }
-bytes = "1.1.0"
 chrono = "0.4.19"
 datafusion = { git = "https://github.com/flock-lab/arrow-datafusion", branch = "flock" }
 driver = { path = "../src/driver" }
 env_logger = "^0.9"
 futures = "0.3"
 humantime = "2.1.0"
-
 itertools = "0.10.0"
 lazy_static = "1.4"
 log = "0.4.14"
