+++ src/driver/Cargo.toml
@@ -15,10 +15,12 @@ arrow = { git = "https://github.com/flock-lab/arrow-rs", branch = "flock", featu
 aws_lambda_events = "0.5"
 base64 = "0.13.0"
 blake2 = "0.9"
+bytes = "1.1.0"
 chrono = "0.4.19"
 daggy = { git = "https://github.com/flock-lab/daggy", branch = "master" }
 datafusion = { git = "https://github.com/flock-lab/arrow-datafusion", branch = "flock" }
 env_logger = "^0.9"
+humantime = "2.1.0"
 lazy_static = "1.4"
 log = "0.4.14"
 runtime = { path = "../../src/runtime" }
@@ -27,11 +29,12 @@ rusoto_iam = "0.47.0"
 rusoto_kinesis = "0.47.0"
 rusoto_lambda = "0.47.0"
 rusoto_logs = "0.47.0"
-tokio = { version = "1.4", features = [ "macros", "io-util", "sync", "rt-multi-thread" ] }
+rusoto_s3 = "0.47.0"
 
 # A list of all of the optional dependencies, some of which are included in the
 # above `features`. They can be opted into by apps.
 serde_json = "1.0"
+tokio = { version = "1.4", features = [ "macros", "io-util", "sync", "rt-multi-thread" ] }
 filetime = { version = "0.2", optional = true }
 fixedbitset = { version = "0.4.0", optional = true }
 glob = { version = "0.3", optional = true }
